+++ molo/core/tests/test_views.py
@@ -3,7 +3,7 @@ import json
 import pytest
 import responses
 
-from datetime import timedelta
+from datetime import timedelta, datetime
 from urlparse import parse_qs
 
 from django.core.files.base import ContentFile
@@ -16,11 +16,15 @@ from molo.core.tests.base import MoloTestCaseMixin
 from molo.core.models import (SiteLanguage, FooterPage,
                               SiteSettings, ArticlePage)
 from molo.core.known_plugins import known_plugins
+from molo.core.tasks import promote_articles
+from molo.core.templatetags.core_tags import \
+    load_descendant_articles_for_section
 
 from mock import patch, Mock
 from six import b
 
 from wagtail.wagtailcore.models import Site
+from wagtail.wagtailimages.tests.utils import Image, get_test_image_file
 from wagtailmedia.models import Media
 
 
@@ -45,6 +49,12 @@ class TestPages(TestCase, MoloTestCaseMixin):
         self.yourmind_sub_fr = self.mk_section_translation(
             self.yourmind_sub, self.french,
             title='Your mind subsection in french')
+
+        # Create an image for running tests on
+        self.image = Image.objects.create(
+            title="Test image",
+            file=get_test_image_file(),
+        )
 
     def test_breadcrumbs(self):
         self.mk_articles(self.yourmind_sub, count=10)
@@ -190,8 +200,9 @@ class TestPages(TestCase, MoloTestCaseMixin):
 
     def test_latest_listing(self):
         en_latest = self.mk_articles(
-            self.yourmind_sub, count=10, featured_in_latest=True)
-
+            self.yourmind_sub, count=10,
+            featured_in_latest_start_date=datetime.now())
+        promote_articles()
         for p in en_latest:
             self.mk_article_translation(
                 p, self.french, title=p.title + ' in french')
@@ -211,9 +222,27 @@ class TestPages(TestCase, MoloTestCaseMixin):
         self.assertNotContains(
             response, 'in french')
 
+    def test_latest(self):
+        en_latest = self.mk_articles(
+            self.yourmind_sub, count=4,
+            featured_in_latest_start_date=datetime.now())
+        promote_articles()
+        for p in en_latest:
+            self.mk_article_translation(
+                p, self.french, title=p.title + ' in french')
+
+        fr_articles = self.mk_articles(self.yourmind_sub, count=10)
+        for p in fr_articles:
+            self.mk_article_translation(
+                p, self.french, title=p.title + ' in french')
+
+        self.assertEquals(self.main.latest_articles().count(), 4)
+
     def test_latest_listing_in_french(self):
         en_latest = self.mk_articles(
-            self.yourmind_sub, count=10, featured_in_latest=True)
+            self.yourmind_sub, count=10,
+            featured_in_latest_start_date=datetime.now())
+        promote_articles()
 
         for p in en_latest:
             self.mk_article_translation(
@@ -297,11 +326,27 @@ class TestPages(TestCase, MoloTestCaseMixin):
             '<li>aeque <em>saepe albucius</em></li>')
 
     def test_featured_homepage_listing(self):
-        self.mk_article(self.yourmind_sub, featured_in_homepage=True)
+        self.mk_article(
+            self.yourmind_sub, featured_in_homepage_start_date=datetime.now())
+        promote_articles()
         response = self.client.get('/')
         self.assertContains(
             response,
             '<p>Sample page description for 0</p>')
+
+    def test_featured_homepage_listing_draft_articles(self):
+        article = self.mk_article(
+            self.yourmind_sub, featured_in_homepage_start_date=datetime.now())
+        article2 = self.mk_article(
+            self.yourmind_sub, featured_in_homepage_start_date=datetime.now())
+        promote_articles()
+        article2.unpublish()
+        self.assertEquals(ArticlePage.objects.live().count(), 1)
+        featured_in_homepage_articles = load_descendant_articles_for_section(
+            {}, self.yourmind_sub, featured_in_homepage=True)
+        self.assertEquals(featured_in_homepage_articles.count(), 1)
+        self.assertEquals(
+            featured_in_homepage_articles.first().title, article.title)
 
     def test_featured_topic_of_the_day(self):
         promote_date = timezone.now() + timedelta(days=-1)
@@ -317,8 +362,64 @@ class TestPages(TestCase, MoloTestCaseMixin):
             response,
             'Topic of the Day')
 
+    def test_social_media_footer(self):
+        default_site = Site.objects.get(is_default_site=True)
+        setting = SiteSettings.objects.create(site=default_site)
+        setting.social_media_links_on_footer_page = json.dumps([
+            {
+                'type': 'social_media_site',
+                'value': {
+                    'title': 'Social Media Site',
+                    'link': 'www.socialmediasite.com',
+                    'image': 0,
+                }}
+        ])
+        setting.save()
+
+        self.footer = FooterPage(
+            title='Footer Page',
+            slug='footer-page')
+        self.footer_index.add_child(instance=self.footer)
+
+        response = self.client.get('/')
+        self.assertContains(
+            response,
+            'www.socialmediasite.com')
+
+    def test_social_media_facebook_sharing(self):
+        default_site = Site.objects.get(is_default_site=True)
+        setting = SiteSettings.objects.create(site=default_site)
+        setting.facebook_sharing = True
+        setting.facebook_image = self.image
+        setting.save()
+
+        self.mk_articles(self.yourmind_sub, count=10)
+
+        response = self.client.get(
+            '/sections/your-mind/your-mind-subsection/test-page-1/')
+        self.assertContains(
+            response,
+            'href="http://www.facebook.com/sharer.php?u=http')
+
+    def test_social_media_twitter_sharing(self):
+        default_site = Site.objects.get(is_default_site=True)
+        setting = SiteSettings.objects.create(site=default_site)
+        setting.twitter_sharing = True
+        setting.twitter_image = self.image
+        setting.save()
+
+        self.mk_articles(self.yourmind_sub, count=10)
+
+        response = self.client.get(
+            '/sections/your-mind/your-mind-subsection/test-page-1/')
+        self.assertContains(
+            response,
+            'href="https://twitter.com/share?url=http')
+
     def test_featured_homepage_listing_in_french(self):
-        en_page = self.mk_article(self.yourmind_sub, featured_in_homepage=True)
+        en_page = self.mk_article(
+            self.yourmind_sub, featured_in_homepage_start_date=datetime.now())
+        promote_articles()
         fr_page = self.mk_article_translation(
             en_page, self.french,
             title=en_page.title + ' in french',